reworking file directories deletedata for modular structure
